requirements.txt
# Core Dependencies - REQUIRED FOR BASIC FUNCTIONALITY
pandas>=1.5.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0

# Authentication - REQUIRED
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0

# Vector RAG System - REQUIRED FOR PERSONALIZED AI
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Machine Learning - REQUIRED FOR FORECASTING
scikit-learn>=1.3.0
statsmodels>=0.14.0
scipy>=1.10.0

# Advanced AI Features - OPTIONAL
transformers>=4.35.0
torch>=2.0.0
xgboost>=1.7.0

# Visualization & Dashboard - OPTIONAL
streamlit>=1.28.0
plotly>=5.15.0
yfinance>=0.2.0

# Optional Enterprise Features
websockets>=11.0.0
psutil>=5.9.0
prometheus-client>=0.18.0

# Optional Advanced Analytics
lifelines>=0.27.0
optuna>=3.4.0
shap>=0.43.0